requirements.txt
playwright>=1.43.0
openai>=1.30.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
